Apply Qodo /improve pass 3: coalesce notifications + widen mask type
1. ulTaskNotifyTake(pdTRUE) clears all pending notifications at once
   so the EOS task runs once per wake, not N times for N queued EOS
   fires. Prevents redundant loops when multiple dedicated-module
   completions pile up before the task runs.

2. Local mask variable widened from uint8_t to uint32_t to match
   gType1EnabledMask type. T1 channels are always 0-4 on NQ1 so
   no functional change, but eliminates a latent truncation bug if
   a future board variant uses higher channel IDs.

SKIPPED: Use xPortIsInsideInterrupt() (4/10) — this API may not
exist on PIC32 MIPS port. uxInterruptNesting is the established
pattern in this codebase (Logger.h uses it).

Entire-Checkpoint: 2c975fff64ea

+++ firmware/src/HAL/ADC/MC12bADC.c
@@ -158,7 +158,7 @@ bool MC12b_WriteStateAll(
     // Do NOT enable CH3 result interrupt — #292 moves T1 result reads to
     // the EOS deferred task, eliminating the ~5μs ISR entry/exit overhead
     // that was the T1 throughput bottleneck.
-    uint8_t mask = 0;
+    uint32_t mask = 0;
     for (i = 0; i < channelConfig->Size; ++i) {
         if (channelConfig->Data[i].Type == AIn_MC12bADC &&
             channelConfig->Data[i].Config.MC12b.ChannelType == 1) {
